plugtrack/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
